+++ libs/ilink-console-tools/tests/e2e/consul-to-env.spec.ts
@@ -21,7 +21,7 @@ describe('ConsulToEnv (e2e)', () => {
       )
       .withEnv('CONSUL_HTTP_TOKEN', CONSUL_HTTP_TOKEN)
       .start();
-    await new Promise((resolve) => setTimeout(resolve, 30000));
+    await new Promise((resolve) => setTimeout(resolve, 3000));
   });
 
   afterAll(async () => {